requirements.txt
aiogram==3.26.0
beautifulsoup4==4.14.3
curl_cffi==0.14.0
python-dotenv==1.2.2
PyYAML==6.0.3
